feat(controller): improve category response handling

+++ src/main/java/com/KambaFlix/Controller/CategoryController.java
@@ -5,6 +5,8 @@ import com.KambaFlix.Controller.response.CategoryResponse;
 import com.KambaFlix.Entity.Category;
 import com.KambaFlix.Service.CategoryService;
 import com.KambaFlix.mapper.CategoryMapper;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -20,29 +22,39 @@ public class CategoryController {
     }
 
     @GetMapping()
-    public List<CategoryResponse> getAllCategory(){
-        List<Category> categoryList = categoryService.findAll();
-        return categoryList.stream()
+    public ResponseEntity<List<CategoryResponse>> getAllCategory(){
+        List<CategoryResponse> categoryList = categoryService.findAll()
+                .stream()
                 .map(CategoryMapper::toCategoryResponce)
                 .toList();
+        return ResponseEntity.ok(categoryList);
     }
 
     @PostMapping("/cadastrar")
-    public CategoryResponse postCadastroDeCategory(@RequestBody CategoryRequest request){
+    public ResponseEntity<CategoryResponse> postCadastroDeCategory(@RequestBody CategoryRequest request){
         Category category = CategoryMapper.toCategory(request);
         Category categorysave = categoryService.cadastroDeCategory(category);
-        return CategoryMapper.toCategoryResponce(categorysave);
+        return ResponseEntity.status(HttpStatus.CREATED).body(CategoryMapper.toCategoryResponce(categorysave));
     }
 
     @GetMapping("/{id}")
-    public CategoryResponse getByCategoryId(@PathVariable Long id){
-        Category category = categoryService.findById(id);
-        return CategoryMapper.toCategoryResponce(category);
+    public ResponseEntity<?> getByCategoryId(@PathVariable Long id){
+        if (categoryService.findById(id) != null) {
+            Category category = categoryService.findById(id);
+            return ResponseEntity.ok(CategoryMapper.toCategoryResponce(category));
+        }else {
+            return  ResponseEntity.status(HttpStatus.NOT_FOUND).body("A categoria com IDs: "+id+" nao encontrado nos nossos banco de dados");
+        }
     }
 
     @DeleteMapping("/{id}")
-    public  void deleteByCategoryId(@PathVariable Long id){
-         categoryService.delete(id);
+    public  ResponseEntity<String> deleteByCategoryId(@PathVariable Long id){
+        if (categoryService.findById(id) != null) {
+            categoryService.delete(id);
+            return ResponseEntity.ok("Categoria com IDs: "+id+" excluido com sucesso");
+        }else {
+            return  ResponseEntity.status(HttpStatus.NOT_FOUND).body("A categoria com IDs: "+id+" nao encontrado nos nossos banco de dados");
+        }
     }
 
 }
